Move testing_helper
