+++ costmap_2d/test/obstacle_tests.cpp
@@ -35,7 +35,7 @@
 #include <costmap_2d/costmap_2d.h>
 #include <costmap_2d/layered_costmap.h>
 #include <costmap_2d/observation_buffer.h>
-#include <testing_helper.h>
+#include <costmap_2d/testing_helper.h>
 #include <set>
 #include <gtest/gtest.h>
 #include <tf/transform_listener.h>
